Use an array for USB endpoint control registers
We may represent ENDPTCTRL[1-7] as an array. This
ends up having very little impact on the RAL output.
The change might help others who want to use these
patched SVDs in another context.

For devices with a single USB peripheral, we change
the peripheral name to 'USB1.' The approach lets us
define a single patch for all the devices. The approach,
however, may lead people to think there's a 'USB2'
somewhere, which might be confusing.
