+++ imxrt-ral/devices/imxrt1021.yaml
@@ -1,5 +1,10 @@
 _svd: "../svd/imxrt1021.svd"
 
+_modify:
+  USB:
+    name: USB1
+
 _include: 
   - "common.yaml"
   - "common_patches/pwm1/submodule_cluster.yaml"
+  - "common_patches/usb1/endptctrl_array.yaml"
