+++ imxrt-ral/devices/imxrt1052.yaml
@@ -3,3 +3,4 @@ _svd: "../svd/imxrt1052.svd"
 _include: 
   - "common.yaml"
   - "common_patches/pwm1/submodule_cluster.yaml"
+  - "common_patches/usb1/endptctrl_array.yaml"
