+++ imxrt-ral/src/imxrt102/imxrt1021/usb1.rs
@@ -1,8 +1,6 @@
 #![allow(non_snake_case, non_upper_case_globals)]
 #![allow(non_camel_case_types)]
 //! USB
-//!
-//! Used by: imxrt1011, imxrt1015
 
 use crate::{RORegister, RWRegister};
 #[cfg(not(feature = "nosync"))]
@@ -930,6 +928,20 @@ pub mod USBCMD {
         pub mod RW {}
     }
 
+    /// Add dTD TripWire - Read/Write
+    pub mod ATDTW {
+        /// Offset (12 bits)
+        pub const offset: u32 = 12;
+        /// Mask (1 bit: 1 << 12)
+        pub const mask: u32 = 1 << offset;
+        /// Read-only values (empty)
+        pub mod R {}
+        /// Write-only values (empty)
+        pub mod W {}
+        /// Read-write values (empty)
+        pub mod RW {}
+    }
+
     /// Setup TripWire - Read/Write
     pub mod SUTW {
         /// Offset (13 bits)
@@ -944,21 +956,7 @@ pub mod USBCMD {
         pub mod RW {}
     }
 
-    /// Add dTD TripWire - Read/Write
-    pub mod ATDTW {
-        /// Offset (14 bits)
-        pub const offset: u32 = 14;
-        /// Mask (1 bit: 1 << 14)
-        pub const mask: u32 = 1 << offset;
-        /// Read-only values (empty)
-        pub mod R {}
-        /// Write-only values (empty)
-        pub mod W {}
-        /// Read-write values (empty)
-        pub mod RW {}
-    }
-
-    /// Frame List Size - (Read/Write or Read Only)
+    /// See also bits 3-2 Frame List Size - (Read/Write or Read Only)
     pub mod FS_2 {
         /// Offset (15 bits)
         pub const offset: u32 = 15;
@@ -2899,7 +2897,7 @@ pub mod ENDPTCTRL0 {
     }
 }
 
-/// Endpoint Control 1
+/// Endpoint Control
 pub mod ENDPTCTRL1 {
 
     /// RX Endpoint Stall - Read/Write 0 End Point OK
@@ -3071,7 +3069,7 @@ pub mod ENDPTCTRL1 {
     }
 }
 
-/// Endpoint Control 2
+/// Endpoint Control
 pub mod ENDPTCTRL2 {
     pub use super::ENDPTCTRL1::RXD;
     pub use super::ENDPTCTRL1::RXE;
@@ -3087,7 +3085,7 @@ pub mod ENDPTCTRL2 {
     pub use super::ENDPTCTRL1::TXT;
 }
 
-/// Endpoint Control 3
+/// Endpoint Control
 pub mod ENDPTCTRL3 {
     pub use super::ENDPTCTRL1::RXD;
     pub use super::ENDPTCTRL1::RXE;
@@ -3103,7 +3101,7 @@ pub mod ENDPTCTRL3 {
     pub use super::ENDPTCTRL1::TXT;
 }
 
-/// Endpoint Control 4
+/// Endpoint Control
 pub mod ENDPTCTRL4 {
     pub use super::ENDPTCTRL1::RXD;
     pub use super::ENDPTCTRL1::RXE;
@@ -3119,7 +3117,7 @@ pub mod ENDPTCTRL4 {
     pub use super::ENDPTCTRL1::TXT;
 }
 
-/// Endpoint Control 5
+/// Endpoint Control
 pub mod ENDPTCTRL5 {
     pub use super::ENDPTCTRL1::RXD;
     pub use super::ENDPTCTRL1::RXE;
@@ -3135,7 +3133,7 @@ pub mod ENDPTCTRL5 {
     pub use super::ENDPTCTRL1::TXT;
 }
 
-/// Endpoint Control 6
+/// Endpoint Control
 pub mod ENDPTCTRL6 {
     pub use super::ENDPTCTRL1::RXD;
     pub use super::ENDPTCTRL1::RXE;
@@ -3151,7 +3149,7 @@ pub mod ENDPTCTRL6 {
     pub use super::ENDPTCTRL1::TXT;
 }
 
-/// Endpoint Control 7
+/// Endpoint Control
 pub mod ENDPTCTRL7 {
     pub use super::ENDPTCTRL1::RXD;
     pub use super::ENDPTCTRL1::RXE;
@@ -3302,25 +3300,25 @@ pub struct RegisterBlock {
     /// Endpoint Control0
     pub ENDPTCTRL0: RWRegister<u32>,
 
-    /// Endpoint Control 1
+    /// Endpoint Control
     pub ENDPTCTRL1: RWRegister<u32>,
 
-    /// Endpoint Control 2
+    /// Endpoint Control
     pub ENDPTCTRL2: RWRegister<u32>,
 
-    /// Endpoint Control 3
+    /// Endpoint Control
     pub ENDPTCTRL3: RWRegister<u32>,
 
-    /// Endpoint Control 4
+    /// Endpoint Control
     pub ENDPTCTRL4: RWRegister<u32>,
 
-    /// Endpoint Control 5
+    /// Endpoint Control
     pub ENDPTCTRL5: RWRegister<u32>,
 
-    /// Endpoint Control 6
+    /// Endpoint Control
     pub ENDPTCTRL6: RWRegister<u32>,
 
-    /// Endpoint Control 7
+    /// Endpoint Control
     pub ENDPTCTRL7: RWRegister<u32>,
 }
 pub struct ResetValues {
@@ -3384,3 +3382,137 @@ impl ::core::ops::Deref for Instance {
 }
 #[cfg(feature = "rtfm")]
 unsafe impl Send for Instance {}
+
+/// Access functions for the USB1 peripheral instance
+pub mod USB1 {
+    use super::ResetValues;
+
+    #[cfg(not(feature = "nosync"))]
+    use super::Instance;
+
+    #[cfg(not(feature = "nosync"))]
+    const INSTANCE: Instance = Instance {
+        addr: 0x402e0000,
+        _marker: ::core::marker::PhantomData,
+    };
+
+    /// Reset values for each field in USB1
+    pub const reset: ResetValues = ResetValues {
+        ID: 0xE4A1FA05,
+        HWGENERAL: 0x00000035,
+        HWHOST: 0x10020001,
+        HWDEVICE: 0x00000011,
+        HWTXBUF: 0x80080B08,
+        HWRXBUF: 0x00000808,
+        GPTIMER0LD: 0x00000000,
+        GPTIMER0CTRL: 0x00000000,
+        GPTIMER1LD: 0x00000000,
+        GPTIMER1CTRL: 0x00000000,
+        SBUSCFG: 0x00000002,
+        CAPLENGTH: 0x00000040,
+        HCIVERSION: 0x00000100,
+        HCSPARAMS: 0x00010011,
+        HCCPARAMS: 0x00000006,
+        DCIVERSION: 0x00000001,
+        DCCPARAMS: 0x00000188,
+        USBCMD: 0x00080000,
+        USBSTS: 0x00000000,
+        USBINTR: 0x00000000,
+        FRINDEX: 0x00000000,
+        DEVICEADDR: 0x00000000,
+        ASYNCLISTADDR: 0x00000000,
+        BURSTSIZE: 0x00000808,
+        TXFILLTUNING: 0x00000000,
+        ENDPTNAK: 0x00000000,
+        ENDPTNAKEN: 0x00000000,
+        CONFIGFLAG: 0x00000001,
+        PORTSC1: 0x10000000,
+        OTGSC: 0x00001120,
+        USBMODE: 0x00005000,
+        ENDPTSETUPSTAT: 0x00000000,
+        ENDPTPRIME: 0x00000000,
+        ENDPTFLUSH: 0x00000000,
+        ENDPTSTAT: 0x00000000,
+        ENDPTCOMPLETE: 0x00000000,
+        ENDPTCTRL0: 0x00800080,
+        ENDPTCTRL1: 0x00000000,
+        ENDPTCTRL2: 0x00000000,
+        ENDPTCTRL3: 0x00000000,
+        ENDPTCTRL4: 0x00000000,
+        ENDPTCTRL5: 0x00000000,
+        ENDPTCTRL6: 0x00000000,
+        ENDPTCTRL7: 0x00000000,
+    };
+
+    #[cfg(not(feature = "nosync"))]
+    #[allow(renamed_and_removed_lints)]
+    #[allow(private_no_mangle_statics)]
+    #[no_mangle]
+    static mut USB1_TAKEN: bool = false;
+
+    /// Safe access to USB1
+    ///
+    /// This function returns `Some(Instance)` if this instance is not
+    /// currently taken, and `None` if it is. This ensures that if you
+    /// do get `Some(Instance)`, you are ensured unique access to
+    /// the peripheral and there cannot be data races (unless other
+    /// code uses `unsafe`, of course). You can then pass the
+    /// `Instance` around to other functions as required. When you're
+    /// done with it, you can call `release(instance)` to return it.
+    ///
+    /// `Instance` itself dereferences to a `RegisterBlock`, which
+    /// provides access to the peripheral's registers.
+    #[cfg(not(feature = "nosync"))]
+    #[inline]
+    pub fn take() -> Option<Instance> {
+        external_cortex_m::interrupt::free(|_| unsafe {
+            if USB1_TAKEN {
+                None
+            } else {
+                USB1_TAKEN = true;
+                Some(INSTANCE)
+            }
+        })
+    }
+
+    /// Release exclusive access to USB1
+    ///
+    /// This function allows you to return an `Instance` so that it
+    /// is available to `take()` again. This function will panic if
+    /// you return a different `Instance` or if this instance is not
+    /// already taken.
+    #[cfg(not(feature = "nosync"))]
+    #[inline]
+    pub fn release(inst: Instance) {
+        external_cortex_m::interrupt::free(|_| unsafe {
+            if USB1_TAKEN && inst.addr == INSTANCE.addr {
+                USB1_TAKEN = false;
+            } else {
+                panic!("Released a peripheral which was not taken");
+            }
+        });
+    }
+
+    /// Unsafely steal USB1
+    ///
+    /// This function is similar to take() but forcibly takes the
+    /// Instance, marking it as taken irregardless of its previous
+    /// state.
+    #[cfg(not(feature = "nosync"))]
+    #[inline]
+    pub unsafe fn steal() -> Instance {
+        USB1_TAKEN = true;
+        INSTANCE
+    }
+}
+
+/// Raw pointer to USB1
+///
+/// Dereferencing this is unsafe because you are not ensured unique
+/// access to the peripheral, so you may encounter data races with
+/// other users of this peripheral. It is up to you to ensure you
+/// will not cause data races.
+///
+/// This constant is provided for ease of use in unsafe code: you can
+/// simply call for example `write_reg!(gpio, GPIOA, ODR, 1);`.
+pub const USB1: *const RegisterBlock = 0x402e0000 as *const _;
