+++ imxrt-ral/src/imxrt106/peripherals/usb.rs
@@ -2899,7 +2899,7 @@ pub mod ENDPTCTRL0 {
     }
 }
 
-/// Endpoint Control 1
+/// Endpoint Control
 pub mod ENDPTCTRL1 {
 
     /// RX Endpoint Stall - Read/Write 0 End Point OK
@@ -3071,7 +3071,7 @@ pub mod ENDPTCTRL1 {
     }
 }
 
-/// Endpoint Control 2
+/// Endpoint Control
 pub mod ENDPTCTRL2 {
     pub use super::ENDPTCTRL1::RXD;
     pub use super::ENDPTCTRL1::RXE;
@@ -3087,7 +3087,7 @@ pub mod ENDPTCTRL2 {
     pub use super::ENDPTCTRL1::TXT;
 }
 
-/// Endpoint Control 3
+/// Endpoint Control
 pub mod ENDPTCTRL3 {
     pub use super::ENDPTCTRL1::RXD;
     pub use super::ENDPTCTRL1::RXE;
@@ -3103,7 +3103,7 @@ pub mod ENDPTCTRL3 {
     pub use super::ENDPTCTRL1::TXT;
 }
 
-/// Endpoint Control 4
+/// Endpoint Control
 pub mod ENDPTCTRL4 {
     pub use super::ENDPTCTRL1::RXD;
     pub use super::ENDPTCTRL1::RXE;
@@ -3119,7 +3119,7 @@ pub mod ENDPTCTRL4 {
     pub use super::ENDPTCTRL1::TXT;
 }
 
-/// Endpoint Control 5
+/// Endpoint Control
 pub mod ENDPTCTRL5 {
     pub use super::ENDPTCTRL1::RXD;
     pub use super::ENDPTCTRL1::RXE;
@@ -3135,7 +3135,7 @@ pub mod ENDPTCTRL5 {
     pub use super::ENDPTCTRL1::TXT;
 }
 
-/// Endpoint Control 6
+/// Endpoint Control
 pub mod ENDPTCTRL6 {
     pub use super::ENDPTCTRL1::RXD;
     pub use super::ENDPTCTRL1::RXE;
@@ -3151,7 +3151,7 @@ pub mod ENDPTCTRL6 {
     pub use super::ENDPTCTRL1::TXT;
 }
 
-/// Endpoint Control 7
+/// Endpoint Control
 pub mod ENDPTCTRL7 {
     pub use super::ENDPTCTRL1::RXD;
     pub use super::ENDPTCTRL1::RXE;
@@ -3302,25 +3302,25 @@ pub struct RegisterBlock {
     /// Endpoint Control0
     pub ENDPTCTRL0: RWRegister<u32>,
 
-    /// Endpoint Control 1
+    /// Endpoint Control
     pub ENDPTCTRL1: RWRegister<u32>,
 
-    /// Endpoint Control 2
+    /// Endpoint Control
     pub ENDPTCTRL2: RWRegister<u32>,
 
-    /// Endpoint Control 3
+    /// Endpoint Control
     pub ENDPTCTRL3: RWRegister<u32>,
 
-    /// Endpoint Control 4
+    /// Endpoint Control
     pub ENDPTCTRL4: RWRegister<u32>,
 
-    /// Endpoint Control 5
+    /// Endpoint Control
     pub ENDPTCTRL5: RWRegister<u32>,
 
-    /// Endpoint Control 6
+    /// Endpoint Control
     pub ENDPTCTRL6: RWRegister<u32>,
 
-    /// Endpoint Control 7
+    /// Endpoint Control
     pub ENDPTCTRL7: RWRegister<u32>,
 }
 pub struct ResetValues {
